+++ backend/controllers/controller_user_test.go
@@ -1,4 +1,4 @@
-package handler
+package controllers
 
 import (
 	"bytes"
