+++ backend/controllers/controller_workspace_test.go
@@ -1,4 +1,4 @@
-package handler
+package controllers
 
 import (
 	"bytes"
