+++ backend/controllers/setup.go
@@ -1,4 +1,4 @@
-package handler
+package controllers
 
 import (
 	"fmt"
