+++ backend/controllers/workspace.go
@@ -1,4 +1,4 @@
-package handler
+package controllers
 
 import (
 	"net/http"
